# docs/swagger.yaml
basePath: /
definitions:
  cicada_web-service-gin_req-res-models.PlayerScore:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      gameid:
        type: string
      id:
        type: string
      name:
        type: string
      score:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.AllWinners:
    properties:
      winnners:
        items:
          $ref: '#/definitions/models.Winner'
        type: array
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.GameRequest:
    properties:
      playerScore:
        items:
          properties:
            id:
              type: string
            name:
              type: string
            score:
              type: integer
          type: object
        type: array
      players:
        items:
          properties:
            amount:
              type: number
            id:
              type: string
            name:
              type: string
          type: object
        type: array
    type: object
...
  models.PlayerScoreResponse:
    properties:
      gameid:
        type: string
      playersScore:
        items:
          $ref: '#/definitions/cicada_web-service-gin_req-res-models.PlayerScore'
        type: array
    type: object
...
  models.PlayersScore:
    properties:
      gameid:
        type: string
      playersScore:
        items:
          $ref: '#/definitions/cicada_web-service-gin_req-res-models.PlayerScore'
        type: array
    type: object
...
  models.Winner:
    properties:
      id:
        type: string
      name:
        type: string
      not_shared_amount:
        type: number
      pools_won_id:
        items:
          type: string
        type: array
      shared_amount:
        type: number
      total_amount:
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: cicada REST API description
  title: cicada REST API
  version: "2.0"
paths:
...
  /play/get-winners/{gameid}:
    get:
      consumes:
      - application/json
      description: get winners with their won amount based on the provided gameid.
      parameters:
      - description: gameid
        in: path
        name: gameid
        required: true
        type: string
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllWinners'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: get winners with their won amount  by gameid
      tags:
      - Play
  /play/winners:
    post:
      consumes:
      - application/json
      description: identify winners from  the given player using thier scores.
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Player data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.GameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllWinners'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: identify winners from  the given player using thier scores.
      tags:
      - Play
  /player/create-player-score:
    post:
      consumes:
      - application/json
      description: Registers players with their scores.
      parameters:
      - description: Bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: Player score data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.PlayersScore'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PlayerScoreResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: Registers players with their scores.
      tags:
      - Player Scores
  /player/get-player-scores/{gameid}:
    get:
      consumes:
      - application/json
      description: Retrieves player scores based on the provided game ID.
      parameters:
      - description: Game ID
        in: path
        name: gameid
        required: true
        type: string
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PlayerScoreResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: Get player scores by game ID
      tags:
      - Player Scores
  /player/update-player-score/{id}:
    put:
      consumes:
      - application/json
      description: Update player score based on the provided ID.
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/cicada_web-service-gin_req-res-models.PlayerScore'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cicada_web-service-gin_req-res-models.PlayerScore'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: Update player score by ID
      tags:
      - Player Scores
  /player/update-player/{gameid}:
    put:
      consumes:
      - application/json
      description: Update player scores based on the provided gameid.
      parameters:
      - description: gameid
        in: path
        name: gameid
        required: true
        type: string
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/cicada_web-service-gin_req-res-models.PlayerScore'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PlayersScore'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - bearer: []
      summary: Update PlayersScore by gameid
      tags:
      - Player Scores
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"